refactor: made code modular
